add board.forms & move posts view
